Remove `Verifier` and `StateVerifier` class and flatten structure
