+++ apps/core/verification_queue.py
@@ -4,7 +4,7 @@ from functools import partial
 from types import FunctionType
 from typing import Optional, Type, Dict, Tuple
 
-from golem.verificator.verifier import Verifier
+from golem.verificator.core_verifier import CoreVerifier
 from twisted.internet.defer import Deferred, gatherResults
 
 from apps.core.verification_task import VerificationTask
@@ -29,7 +29,7 @@ class VerificationQueue:
         self._paused = False
 
     def submit(self,
-               verifier_class: Type[Verifier],
+               verifier_class: Type[CoreVerifier],
                subtask_id: str,
                deadline: int,
                cb: FunctionType,
@@ -65,14 +65,15 @@ class VerificationQueue:
             if entry and verifier_cls:
                 self._run(entry, verifier_cls)
 
-    def _next(self) -> Tuple[Optional[VerificationTask], Optional[Verifier]]:
+    def _next(self) -> Tuple[Optional[VerificationTask],
+                             Optional[CoreVerifier]]:
         try:
             return self._queue.get(block=False)
         except queue.Empty:
             return None, None
 
     def _run(self, entry: VerificationTask,
-             verifier_cls: Type[Verifier]) -> None:
+             verifier_cls: Type[CoreVerifier]) -> None:
         subtask_id = entry.subtask_id
 
         logger.info("Running verification of subtask %r", subtask_id)
