+++ tests/golem/verificator/test_verifier.py
@@ -3,7 +3,8 @@ from unittest import TestCase
 
 from freezegun import freeze_time
 
-from golem.verificator.verifier import StateVerifier, SubtaskVerificationState
+from golem.verificator.constants import SubtaskVerificationState
+from golem.verificator.core_verifier import CoreVerifier
 
 
 @freeze_time()
@@ -21,6 +22,6 @@ class VerifierTest(TestCase):
             assert kwargs['result']['time_started'] == time
             assert kwargs['result']['time_ended'] == time
 
-        sv = StateVerifier()
+        sv = CoreVerifier()
         sv.callback = callback
         sv.task_timeout(subtask_id)
